+++ libft/ft_strlcpy.c
@@ -6,13 +6,13 @@
 /*   By: rida-cos <rida-cos@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/19 13:50:40 by rida-cos          #+#    #+#             */
-/*   Updated: 2025/07/20 18:25:34 by rida-cos         ###   ########.fr       */
+/*   Updated: 2025/08/04 18:25:54 by rida-cos         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlcpy(char *dst, const char *src, size_t size)
+size_t	ft_strlcpy(char *dst, const char *src, size_t size)
 {
 	size_t	len;
 	size_t	i;
