+++ libft/ft_strtrim.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strtrim.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: rdcm <rdcm@student.42.fr>                  +#+  +:+       +#+        */
+/*   By: rida-cos <rida-cos@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/20 14:29:24 by rida-cos          #+#    #+#             */
-/*   Updated: 2025/07/30 00:14:12 by rdcm             ###   ########.fr       */
+/*   Updated: 2025/08/04 19:21:56 by rida-cos         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,9 +28,9 @@ static int	is_in_set(char c, const char *set)
 
 char	*ft_strtrim(char const *s1, char const *set)
 {
-	int		start;
-	int		end;
-	int		trimmed_len;
+	size_t	start;
+	size_t	end;
+	size_t	trimmed_len;
 	char	*trimmed;
 
 	if (!s1 || !set)
